Add token-based auth on all POST/PUT/PATCH/DELETE endpoints

+++ packages/express-backend/src/server.js
@@ -48,16 +48,38 @@ app.get('/users/:id', async (req, res) => {
 
 app.delete('/users/:userId', async (req, res) => {
   const userId = req.params.userId
-  await Database.deleteUser(userId)
 
-  res.status(204).send()
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      await Database.deleteUser(userId)
+      res.status(204).send()
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.delete('/items/:itemId', async (req, res) => {
   const itemId = req.params.itemId
-  await Database.deleteItem(itemId)
+  const item = Database.findItemById(itemId)
 
-  res.status(204).send()
+  // If the item does not exist, return success
+  if (item === undefined) {
+    res.status(204).send()
+    return
+  }
+
+  const userId = item.seller_id
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      await Database.deleteItem(itemId)
+      res.status(204).send()
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.get('/items', async (req, res) => {
@@ -91,12 +113,20 @@ app.get('/users/:userId/wishlist', async (req, res) => {
 
 app.put('/users/:userId/wishlist/:itemId', async (req, res) => {
   const userId = req.params.userId
-  const itemId = req.params.itemId
 
-  const editedWishlist = await Database.addToWishList(userId, itemId)
-  const statusCode = editedWishlist === undefined ? 500 : 200
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      const itemId = req.params.itemId
 
-  res.status(statusCode).send(editedWishlist)
+      const editedWishlist = await Database.addToWishList(userId, itemId)
+      const statusCode = editedWishlist === undefined ? 500 : 200
+
+      res.status(statusCode).send(editedWishlist)
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.delete('/users/:userId/wishlist/:itemId', async (req, res) => {
@@ -123,22 +153,16 @@ app.post('/login', async (req, res) => {
 app.post('/users/:userId/items', async (req, res) => {
   const userId = req.params.userId
 
-  const authPayload = Passwords.extractToken(req, res)
-  if (authPayload === undefined) {
-    return
-  }
-
-  const { username: authUsername, userId: authUserId } = authPayload
-  console.log(`Request from ${authUsername} (${authUserId})`)
-  if (userId !== authUserId) {
-    res.status(401).end()
-    return
-  }
-
-  const createdItem = await Database.createItem(req.body, userId)
-  const statusCode = createdItem === undefined ? 500 : 201
-
-  res.status(statusCode).send(createdItem || {})
+  Passwords.authenticateUser(req, userId)
+    .then(async () => {
+      const createdItem = await Database.createItem(req.body, userId)
+      const statusCode = createdItem === undefined ? 500 : 201
+      res.status(statusCode).send(createdItem || {})
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.get('/users/:userId/items', async (req, res) => {
@@ -158,22 +182,31 @@ app.get('/users/:userId/items-bought', async (req, res) => {
 })
 
 app.patch('/users/:sellerId/items/:itemId', async (req, res) => {
-  const itemId = req.params.itemId
-  const buyerId = req.body.buyerId
+  const sellerId = req.params.sellerId
 
-  if (!buyerId) {
-    return res.status(400).send({ error: 'buyerId is required' })
-  }
+  Passwords.authenticateUser(req, sellerId)
+    .then(async () => {
+      const itemId = req.params.itemId
+      const buyerId = req.body.buyerId
 
-  try {
-    const updatedItem = await Database.updateItemBuyerAndAddToUserBought(
-      itemId,
-      buyerId
-    )
-    res.status(200).send(updatedItem)
-  } catch (error) {
-    res.status(500).send({ error: error.message })
-  }
+      if (!buyerId) {
+        return res.status(400).send({ error: 'buyerId is required' })
+      }
+
+      try {
+        const updatedItem = await Database.updateItemBuyerAndAddToUserBought(
+          itemId,
+          buyerId
+        )
+        res.status(200).send(updatedItem)
+      } catch (error) {
+        res.status(500).send({ error: error.message })
+      }
+    })
+    .catch((error) => {
+      const { statusCode, message } = error
+      res.status(statusCode).send(message)
+    })
 })
 
 app.listen(process.env.PORT || port, () => {
